Centralize the traffic filter so only relevant traffic is shown.
