+++ views/adsb_top_view_scope.py
@@ -3,232 +3,20 @@ View element for a "radar scope" that looks from the top downwards.
 """
 
 import math
-from datetime import datetime
-from numbers import Number
 from typing import Tuple
-from views.hud_elements import apply_declination
 
 import pygame
-from common_utils import fast_math, local_debug, units
-from common_utils.fast_math import interpolatef
+from common_utils import fast_math, units
 from common_utils.task_timer import TaskProfiler
 from configuration import configuration
+from core_services import zoom_tracker
 from data_sources.ahrs_data import AhrsData
 from data_sources.data_cache import HudDataCache
 from data_sources.traffic import Traffic
 from rendering import colors, drawing
 
 from views.adsb_element import AdsbElement
-
-
-class ZoomTracker:
-    """
-    Tracks what our current zoom is and should be.
-    Handles gracefully transitioning the desired
-    zoom distance, while providing flapping prevention.
-    """
-
-    SECONDS_FOR_ZOOM = 3
-    MINIMUM_SECONDS_BETWEEN_ZOOM_CHANGE = SECONDS_FOR_ZOOM * 5
-
-    def __init__(
-        self,
-        starting_zoom: Tuple[int, int]
-    ) -> None:
-        super().__init__()
-
-        self.__last_changed__ = datetime.utcnow()
-        self.__last_zoom__ = starting_zoom
-        self.__target_zoom__ = starting_zoom
-
-    def set_target_zoom(
-        self,
-        new_target_zoom: Tuple[int, int]
-    ):
-        """
-        Sets the desired target zoom distance.
-
-        If a zoom target has been set too recently
-        then the request is ignored. (Anti-flapping)
-
-        Args:
-            new_target_zoom (Tuple[int, int]): The total distance of the zoom and distance between rings.
-        """
-
-        if new_target_zoom is None:
-            return
-
-        if new_target_zoom[0] == self.__target_zoom__[0]:
-            return
-
-        delta_since_last_change = (datetime.utcnow() - self.__last_changed__).total_seconds()
-
-        if delta_since_last_change < ZoomTracker.MINIMUM_SECONDS_BETWEEN_ZOOM_CHANGE:
-            return
-
-        print("Old target zoom={}".format(self.__target_zoom__))
-        print("Setting new target zoom={}".format(new_target_zoom))
-
-        self.__last_zoom__ = self.__target_zoom__
-        self.__last_changed__ = datetime.utcnow()
-        self.__target_zoom__ = new_target_zoom
-
-    def get_target_zoom(
-        self
-    ) -> Tuple[Number, int]:
-        """
-        Get what our ideal, current zoom is.
-
-        Returns a tuple that is the ideal distance
-        AND the distance between rings.
-
-        Returns:
-            [Number, int]: The range and step of the scope rings
-        """
-        delta_since_last_change = (datetime.utcnow() - self.__last_changed__).total_seconds()
-
-        proportion_into_zoom = delta_since_last_change / ZoomTracker.SECONDS_FOR_ZOOM
-
-        if proportion_into_zoom >= 1.0:
-            return self.__target_zoom__
-
-        computed_range = interpolatef(
-            self.__last_zoom__[0],
-            self.__target_zoom__[0],
-            proportion_into_zoom)
-
-        # Determine the stepping to use based on
-        # stepping of the rings. We want to always
-        # use the largest stepping.
-        #
-        # We choose this way so the whole process has visual consistency
-        # and does not startle the aviator.
-        previous_zoom_stepping = self.__last_zoom__[1]
-        new_zoom_stepping = self.__target_zoom__[1]
-
-        target_zoom_step = max(new_zoom_stepping, previous_zoom_stepping)
-
-        return [computed_range, target_zoom_step]
-
-
-DEFAULT_SCOPE_RANGE = (10, 5)
-
-SCOPE_RANGES = [
-    (1, 0),
-    (2, 1),
-    (5, 3),
-    (10, 5),
-    (15, 5),
-    (20, 10),
-    (50, 25)]
-
-
-def __get_maximum_scope_range__() -> Tuple[int, int]:
-    """
-    Get the maximum
-
-    Returns:
-        (int, int): Tuple of the maximum distance for the scope, and the distance between each ring.
-    """
-    return SCOPE_RANGES[-1]
-
-
-# A value of "12" is how far you will fly in
-# ten minutes. Making it 12 x 2 will range
-# the scope based on how far you will
-# fly in 5 minutes.
-__DISTANCE_PREDICTION_SCALER__ = 12 * 2
-
-
-def __get_ideal_scope_range__(
-    groundspeed: float
-) -> Tuple[int, int]:
-    """
-    Given a ground speed, figure out how far the scope should be.
-    This is done by figuring out how far you will be in 10 minutes
-
-    Args:
-        groundspeed (float): The speed to calculate the ideal scope range from. This is in final units (so MPH, KPH, KNOTS)
-
-    Returns:
-        (int, int): The maximum distance the scope will cover and the distance between each ring.
-
-    >>> __get_ideal_scope_range__(0.0)
-    (1, 0)
-    >>> __get_ideal_scope_range__(0)
-    (1, 0)
-    >>> __get_ideal_scope_range__(-1 * __DISTANCE_PREDICTION_SCALER__)
-    (1, 0)
-    >>> __get_ideal_scope_range__(-1.0 * __DISTANCE_PREDICTION_SCALER__)
-    (1, 0)
-    >>> __get_ideal_scope_range__(0.5 * __DISTANCE_PREDICTION_SCALER__)
-    (1, 0)
-    >>> __get_ideal_scope_range__(1 * __DISTANCE_PREDICTION_SCALER__)
-    (1, 0)
-    >>> __get_ideal_scope_range__(1.0 * __DISTANCE_PREDICTION_SCALER__)
-    (1, 0)
-    >>> __get_ideal_scope_range__(1.1 * __DISTANCE_PREDICTION_SCALER__)
-    (2, 1)
-    >>> __get_ideal_scope_range__(1.5 * __DISTANCE_PREDICTION_SCALER__)
-    (2, 1)
-    >>> __get_ideal_scope_range__(1.9 * __DISTANCE_PREDICTION_SCALER__)
-    (2, 1)
-    >>> __get_ideal_scope_range__(2.0 * __DISTANCE_PREDICTION_SCALER__)
-    (2, 1)
-    >>> __get_ideal_scope_range__(2.1 * __DISTANCE_PREDICTION_SCALER__)
-    (5, 3)
-    >>> __get_ideal_scope_range__(5 * __DISTANCE_PREDICTION_SCALER__)
-    (5, 3)
-    >>> __get_ideal_scope_range__(9.9 * __DISTANCE_PREDICTION_SCALER__)
-    (10, 5)
-    >>> __get_ideal_scope_range__(10 * __DISTANCE_PREDICTION_SCALER__)
-    (10, 5)
-    >>> __get_ideal_scope_range__(14.9 * __DISTANCE_PREDICTION_SCALER__)
-    (15, 5)
-    >>> __get_ideal_scope_range__(15 * __DISTANCE_PREDICTION_SCALER__)
-    (15, 5)
-    >>> __get_ideal_scope_range__(19.9 * __DISTANCE_PREDICTION_SCALER__)
-    (20, 10)
-    >>> __get_ideal_scope_range__(20 * __DISTANCE_PREDICTION_SCALER__)
-    (20, 10)
-    >>> __get_ideal_scope_range__(20.1 * __DISTANCE_PREDICTION_SCALER__)
-    (50, 25)
-    >>> __get_ideal_scope_range__(50 * __DISTANCE_PREDICTION_SCALER__)
-    (50, 25)
-    >>> __get_ideal_scope_range__(100 * __DISTANCE_PREDICTION_SCALER__)
-    (50, 25)
-    """
-
-    predicted_travel_distance = math.fabs(groundspeed) / __DISTANCE_PREDICTION_SCALER__
-
-    # The idea is to return the first range in the set
-    # that is further than you will fly in 10 minutes.
-    for possible_range in SCOPE_RANGES:
-        range_distance = possible_range[0]
-        if range_distance >= predicted_travel_distance:
-            return possible_range
-
-    return __get_maximum_scope_range__()
-
-
-def __get_groundspeed__(
-    orientation: AhrsData
-) -> float:
-    is_valid_groundspeed = orientation.groundspeed is not None and isinstance(orientation.groundspeed, Number)
-    is_valid_airspeed = orientation.airspeed is not None and isinstance(orientation.airspeed, Number)
-
-    airspeed = units.get_converted_units(
-        configuration.CONFIGURATION.get_units(),
-        orientation.airspeed * units.feet_to_nm) if is_valid_airspeed else 0.0
-
-    groundspeed = units.get_converted_units(
-        configuration.CONFIGURATION.get_units(),
-        orientation.groundspeed * units.yards_to_nm) if is_valid_groundspeed else 0
-
-    if (local_debug.is_debug() or not is_valid_groundspeed) and is_valid_airspeed:
-        return airspeed
-
-    return groundspeed
+from views.hud_elements import apply_declination
 
 
 class AdsbTopViewScope(AdsbElement):
@@ -270,8 +58,6 @@ class AdsbTopViewScope(AdsbElement):
         self.__scope_center__ = [
             self.__center_x__,
             self.__center_y__ + int(self.__center_y__ >> 1)]
-
-        self.__zoom_tracker__ = ZoomTracker(DEFAULT_SCOPE_RANGE)
 
         size = self.__framebuffer_size__[1] * 0.04
         half_size = int((size / 2.0) + 0.5)
@@ -389,11 +175,9 @@ class AdsbTopViewScope(AdsbElement):
         # TODO - Consider tail numbers to the side, with lines
         # that connect the number to the target.
 
-        display_distance = units.get_converted_units(
-            configuration.CONFIGURATION.get_units(),
-            traffic.distance)
+        (is_within_threshold, display_distance) = zoom_tracker.INSTANCE.is_in_inner_range(traffic.distance)
 
-        if display_distance > scope_range:
+        if not is_within_threshold:
             return
 
         pixels_from_center = self.__get_pixel_distance__(
@@ -424,7 +208,7 @@ class AdsbTopViewScope(AdsbElement):
             points = self.__get_traffic_indicator__(
                 [screen_x, screen_y],
                 orientation.get_onscreen_projection_heading(),
-                traffic.track)
+                apply_declination(traffic.track))
             drawing.renderer.polygon(framebuffer, target_color, points, not self.__reduced_visuals__)
         else:
             drawing.renderer.filled_circle(
@@ -623,19 +407,6 @@ class AdsbTopViewScope(AdsbElement):
                 heading_to_draw,
                 scope_range)
 
-    def __get_scope_zoom__(
-        self,
-        orientation: AhrsData
-    ) -> Tuple[Number, float]:
-        groundspeed = __get_groundspeed__(orientation)
-
-        ideal_range = __get_ideal_scope_range__(groundspeed)
-
-        # TODO - Make sure that the zoom rings are properly scoping in and out
-        self.__zoom_tracker__.set_target_zoom(ideal_range)
-
-        return self.__zoom_tracker__.get_target_zoom()
-
     def render(
         self,
         framebuffer: pygame.Surface,
@@ -654,11 +425,13 @@ class AdsbTopViewScope(AdsbElement):
         # TODO: MORE TESTING!!!
 
         with TaskProfiler('views.adsb_top_view_scope.AdsbTopViewScope.setup'):
-            scope_range = self.__get_scope_zoom__(orientation)
+            scope_range = zoom_tracker.INSTANCE.get_target_zoom()
             traffic_reports = HudDataCache.get_reliable_traffic()
             traffic_reports.sort(
                 key=lambda traffic: traffic.distance,
                 reverse=True)
+
+        near_target_distance = zoom_tracker.INSTANCE.get_target_threshold_distance()
 
         with TaskProfiler('views.adsb_top_view_scope.AdsbTopViewScope.render'):
             self.__render_ownship__(framebuffer)
@@ -670,14 +443,14 @@ class AdsbTopViewScope(AdsbElement):
             self.__draw_all_compass_headings__(
                 framebuffer,
                 orientation,
-                scope_range[0])
+                near_target_distance)
 
             # pylint: disable=expression-not-assigned
             [self.__render_on_screen_target__(
                 framebuffer,
                 orientation,
                 traffic,
-                scope_range[0],
+                near_target_distance,
                 first_ring_pixel_radius) for traffic in traffic_reports]
